modifying points

+++ src/commands.rs
@@ -302,7 +302,7 @@ pub async fn handle_duel_command(
         None => "100",
     };
 
-    let points: i32 = match points.parse() {
+    let points: i64 = match points.parse() {
         Result::Ok(p) => match p {
             p if p < 0 => {
                 return messaging::send_duel_err(
@@ -695,7 +695,7 @@ pub async fn handle_gamble_command(
     let mut cmd_iter = msg.text().split(' ');
     cmd_iter.next();
     let wager = match cmd_iter.next() {
-        Some(w) => match w.parse::<i32>() {
+        Some(w) => match w.parse::<i64>() {
             Ok(w) => w,
             Err(_) => {
                 return messaging::reply_to(
@@ -938,7 +938,7 @@ pub async fn handle_setpoints_command(
             None => "100",
         };
 
-        let new_points: i32 = match points.parse() {
+        let new_points: i64 = match points.parse() {
             Result::Ok(p) => match p {
                 p if p < 0 => {
                     return messaging::reply_to(client, msg, "provide a positive value").await;
@@ -1000,7 +1000,7 @@ pub async fn handle_gift_command(
         None => "100",
     };
 
-    let new_points: i32 = match points.parse() {
+    let new_points: i64 = match points.parse() {
         Result::Ok(p) => match p {
             p if p < 0 => {
                 return messaging::reply_to(client, msg, "provide a positive value").await;
